Refactor blocks and hovering toolbar in editor
